add some databundle tests
